Update&Fix vistas,repos,service,entites,dtos,controller

+++ src/main/java/com/oo2/grupo3/models/dtos/responses/HorarioLaboralResponseDTO.java
@@ -19,7 +19,7 @@ import lombok.Setter;
 
 public class HorarioLaboralResponseDTO {
 	
-    private Long idHorarioLaboral;
+    private Integer idHorarioLaboral;
     private DiaSemana diaSemana;
     private LocalTime horaInicio;
     private LocalTime horaFin;
